+++ batch/job/main.tf
@@ -3,7 +3,7 @@ resource "aws_batch_job_definition" "this_job" {
   name                  = each.value.name_effective
   container_properties  = jsonencode(each.value.container_properties)
   parameters            = each.value.parameter_map
-  platform_capabilities = ["EC2"] # TODO
+  platform_capabilities = each.value.platform_capability_list
   propagate_tags        = true
   tags                  = each.value.tags
   type                  = "container"